I improved exception handling and added controller validations

+++ src/main/java/com/capitole/ecommerce/price/service/IPricesService.java
@@ -7,5 +7,5 @@ import java.util.Optional;
 
 public interface IPricesService {
 
-    Optional<Price> getPriceByDateProductAndBrand(ZonedDateTime date, int brandId, int productId);
+    Optional<Price> getPriceByDateProductAndBrand(ZonedDateTime date, long brandId, long productId);
 }
